#7 - adicionado comentários para documentação e ajuste na responsábilidade das classes
