+++ src/main/java/telegrambot/service/BotMessages.java
@@ -1,5 +1,8 @@
 package telegrambot.service;
 
+/**
+ * BotMessages - Classe responsável por salvar mensagens genéricas que são úteis para os 3 idiomas suportados no i18n
+ */
 public class BotMessages {
     public static final String BOT_LANGUAGE = "Qual o seu idioma? \n" +
             "What's your languague? \n" +
